[PE] Fix video overflow clip
Previously video with object-fit:none or object-fit:cover was
not clipped when the contents overflowed the replaced content
box, because it was treated as an image that was supposed
not to paint outside of the replaced content box but not true.

Simplify overflow clip logic of replaced element. Remove the
dependency of image overflow clip on compositing state.

We treat videos as images, so the above change fixes video clip
issue.

ReplacedPainter will emit paint chunk with overflow clip
depending on paint property only. No longer depends on other
conditions. This applies to all replaced elements.

Bug: 912357
Change-Id: I5de102f7b6b9ad79f372392de44012deb8a9cfec