[LayoutNG] Fix SVG replaced element sizes
The old code made an incorrect assumption that if Legacy's
IntrinsicSizingInfo did not compute width/height, the fallback
should be LayoutBox::IntrinsicSize.

I thought I was being clever, and was trying to keep sizing computation
logic out of LayoutBox. Unfortunately, not enough information
is exposed by LayoutBox to compute ReplacedSize from raw
IntrinsicSize, and IntrinsicSizingInfo must be used.

The correct fix is to use ConstraintSpace::AvailableSize if
IntrinsicSizingInfo is unavailable.

Bug: 916387
Change-Id: Ib58bbd4775aa9fe1129bc5613d1652123b12670e